💚 [docs] fix build

+++ scripts/index.ts
@@ -40,7 +40,8 @@ const scriptHandlers = {
                 './docs/.vuepress/node_modules/.bin/vuepress build docs'
             ).toString()
         );
-        fs.rmSync('./dist/docs', { recursive: true });
+        if (fs.existsSync('./dist/docs'))
+            fs.rmSync('./dist/docs', { recursive: true });
         fs.mkdirSync('./dist/docs', { recursive: true });
         fs.cpSync('./docs/.vuepress/dist/', './dist/docs', { recursive: true });
     },
